Sun gcc warnings as errors in 64-bit mode.

+++ groups/bsl/bslmt/bslmt_threadutilimpl_pthread.cpp
@@ -140,7 +140,7 @@ static int initPthreadAttribute(pthread_attr_t                 *destination,
         // Note sometimes PTHREAD_STACK_MIN is a function so cache the call to
         // a variable.
 
-        const int pthreadStackMin = PTHREAD_STACK_MIN;
+        const int pthreadStackMin = static_cast<int>(PTHREAD_STACK_MIN);
         if (stackSize < pthreadStackMin) {
             stackSize = pthreadStackMin;
         }
